Update the MSCK REPAIR TABLE docs
New Athena users that don't come from the Hive/Presto/Spark world don't know about the Hive style partitioning scheme and will attempt to use MSCK REPAIR TABLE on for example structures created by Kinesis Firehose or CloudTrail. To help people discover the limitations of the command, and what the alternatives are a note about Hive style partitioning is added, and also a note and link to the ALTER TABLE ADD PARTITION command.

A note about only using MSCK REPAIR TABLE to load partitions for new tables, or tables where the partitions have changed significantly is also added, because the command is very slow to run as a way to just add a single partition.