Use the same table in both scenarios
The first scenario shows the schema of a table, but the second scenario doesn't. For some reason the second scenario assumes a different table schema, which is confusing.

Changing the second scenario to assume the same table schema as the first scenario makes the examples work better together, and contrast the benefits and downsides of both approaches.

Using a single partition key in the second scenario also shows that you don't have to have a separate partition column per S3 "directory", which is something that isn't at all clear from the example as it is now.

Also change the heading to not imply that the data in the second scenario isn't partitioned. It is partitioned, it's just not partitioned in the Hive style. The heading of the first scenario makes it clear that it describes Hive style partitioning, and the second scenario should make it clearer that it describes an alternative style, not a lack of partitioning (that's something entirely different).