Update json.md
The current documentation for the case.insensitive property is misleading or wrong. Athena does not require the data to have lower case keys, as is implied. It also left out a very important part of how to use the property, that without explicit mappings the properties will not be found.

What happens is that Athena will lower case keys. Column names will always be lower cased when you create the table through Athena (not sure if it's the same if you do it through Glue). This means that when `case.insensitive` is false Athena will look for a lower case key, but the serde will have preserved the casing of the keys, and you end up with NULL for all columns where the underlying key has upper case characters.

With the behaviour just described, the only reason to set this property is to get around duplicate key errors, and I've provided guidance for that in the documentation. By default, if you have the properties "URL" and "Url" you will get a duplicate key error (`HIVE_CURSOR_ERROR: Row is not a valid JSON Object - JSONException: Duplicate key "url"`) because they will both be lower cased to the same string. By setting the property to false and providing mappings you can get around that problem.