Add a query example to the partition projection docs
When reading through the partition projection docs, one thing that wasn't immediately clear to me was that there is no magic going on when the projection type is "date". I was somehow expecting it to allow me to use dates in queries, which would be formatted using the configured format. By adding this example I hope that other people will get how it actually works faster than I did.

In other words, after my first reading I expected this query to be possible: `SELECT * FROM my_table WHERE datehour = DATE '2018-02-03 10:00'`, but in fact, what happens is rather that the values for the partition key `datehour` have to be literally what gets interpolated into the `storage.location.template`.

This change also adds a name to the example table (because that was needed to make the query example connect to the DDL above it), and also lower cases the name of the column in the DDL for consistency (it was upper case in the partition key section, but lower case in the table properties).