Update the DATE and TIMESTAMP data type docs
The `DATE` and `TIMESTAMP` formats are _not_ examples of "UNiX format" (weird and inconsistent capitalization or no). The date format would best be described as ISO standard format, while the timestamp format is (at least according to the OpenCSVSerDe page in the Athena documentation) an example of the format expected by `java.sql.Timestamp` (and consequently Hive, I would assume).

I also made sure that the two entries look more consistent by using the same capitalisation for the format string, and that both have examples.