Use consistent casing of TRUE/FALSE
In text, `true` and `false` were capitalised, while in code examples they were not. This normalises the capitalisation to the variant used in the code examples, do decrease confusion. I assume both capitalisations work, but it makes more sense to use the same exact variant when referring to it in text as is then used in the following code example.