Update querying-r53-resolver-logs.md
Cx ran into this issue with the example being incorrect, updated to the correct variable the cx was able to use to retrieve the info. Not familiar with Athena, so please feel free to adjust as needed but this was the matching field in the resolver logs and the cx was able to use it to pull the info properly.